requirements.txt
chess
pyyaml
pandas
